+++ lib/kitchen/directions/book_answer_key_container/v1.rb
@@ -4,10 +4,11 @@ module Kitchen::Directions::BookAnswerKeyContainer
   class V1
     renderable
 
-    def bake(book:)
+    def bake(book:, solutions_plural: true)
       @metadata = book.metadata.children_to_keep.copy
+      @solutions_or_solution = solutions_plural ? 'solutions' : 'solution'
       book.body.append(child: render(file: 'eob_solutions_container.xhtml.erb'))
-      book.body.first('div.os-eob.os-solutions-container')
+      book.body.first("div.os-eob.os-#{@solutions_or_solution}-container")
     end
   end
 end