+++ lib/kitchen/directions/move_solutions_to_answer_key/main.rb
@@ -3,13 +3,14 @@
 module Kitchen
   module Directions
     module MoveSolutionsToAnswerKey
-      def self.v1(chapter:, metadata_source:, strategy:, append_to:, strategy_options: {})
+      def self.v1(chapter:, metadata_source:, strategy:, append_to:, strategy_options: {}, solutions_plural: true)
         V1.new.bake(
           chapter: chapter,
           metadata_source: metadata_source,
           strategy: strategy,
           append_to: append_to,
-          strategy_options: strategy_options
+          strategy_options: strategy_options,
+          solutions_plural: solutions_plural
         )
       end
     end