+++ lib/kitchen/directions/move_solutions_to_answer_key/v1.rb
@@ -2,7 +2,7 @@
 
 module Kitchen::Directions::MoveSolutionsToAnswerKey
   class V1
-    def bake(chapter:, metadata_source:, strategy:, append_to:, strategy_options: {})
+    def bake(chapter:, metadata_source:, strategy:, append_to:, strategy_options: {}, solutions_plural: true)
       strategy =
         case strategy
         when :calculus
@@ -17,9 +17,11 @@ module Kitchen::Directions::MoveSolutionsToAnswerKey
           raise 'No such strategy'
         end
 
+      solutions_or_solution = solutions_plural ? 'solutions' : 'solution'
       append_to.append(child:
         <<~HTML
-          <div class="os-eob os-solutions-container" data-type="composite-page" data-uuid-key=".solutions#{chapter.count_in(:book)}">
+          <div class="os-eob os-#{solutions_or_solution}-container" data-type="composite-page" \
+          data-uuid-key=".#{solutions_or_solution}#{chapter.count_in(:book)}">
             <h2 data-type="document-title">
               <span class="os-text">#{I18n.t(:chapter)} #{chapter.count_in(:book)}</span>
             </h2>