+++ spec/directions/book_answer_key_container/v1_spec.rb
@@ -6,7 +6,7 @@ RSpec.describe Kitchen::Directions::BookAnswerKeyContainer::V1 do
   let(:book) do
     book_containing(html:
       <<~HTML
-        #{metadata(title: 'Book Title')}
+        #{metadata_element}
         <div data-type="chapter">
           <div data-type="page">
             This is a page
@@ -25,49 +25,34 @@ RSpec.describe Kitchen::Directions::BookAnswerKeyContainer::V1 do
           </h1>
           <div data-type="metadata" style="display: none;">
             <h1 data-type="document-title" itemprop="name">Answer Key</h1>
-            <div class="authors">
-            <span id="author-1_copy_1"><a>OpenStaxCollege</a></span>
-          </div><div class="publishers">
-            <span id="publisher-1_copy_1"><a>OpenStaxCollege</a></span>
-          </div><div class="print-style">
-            <span data-type="print-style">ccap-calculus</span>
-          </div><div class="permissions">
-            <p class="copyright">
-              <span id="copyright-holder-1_copy_1"><a>OSCRiceUniversity</a></span>
-            </p>
-            <p class="license">
-              <a>CC BY</a>
-            </p>
-          </div><div itemprop="about" data-type="subject">Math</div>
+            <div class="authors" id="authors_copy_1">Authors</div>
+            <div class="publishers" id="publishers_copy_1">Publishers</div>
+            <div class="print-style" id="print-style_copy_1">Print Style</div>
+            <div class="permissions" id="permissions_copy_1">Permissions</div>
+            <div data-type="subject" id="subject_copy_1">Subject</div>
           </div>
         </div>
       HTML
     )
   end
 
-  def metadata(title:, id_suffix: '')
-    <<~HTML
-      <div data-type="metadata" style="display: none;">
-        <h1 data-type="document-title" itemprop="name">#{title}</h1>
-        <div class="authors">
-          <span id="author-1#{id_suffix}" ><a>OpenStaxCollege</a></span>
+  it 'generates container with solution (singular) class' do
+    expect(described_class.new.bake(book: book, solutions_plural: false)).to match_normalized_html(
+      <<~HTML
+        <div class="os-eob os-solution-container" data-type="composite-chapter" data-uuid-key=".solution">
+          <h1 data-type="document-title">
+            <span class="os-text">Answer Key</span>
+          </h1>
+          <div data-type="metadata" style="display: none;">
+            <h1 data-type="document-title" itemprop="name">Answer Key</h1>
+            <div class="authors" id="authors_copy_1">Authors</div>
+            <div class="publishers" id="publishers_copy_1">Publishers</div>
+            <div class="print-style" id="print-style_copy_1">Print Style</div>
+            <div class="permissions" id="permissions_copy_1">Permissions</div>
+            <div data-type="subject" id="subject_copy_1">Subject</div>
+          </div>
         </div>
-        <div class="publishers">
-          <span id="publisher-1#{id_suffix}"><a>OpenStaxCollege</a></span>
-        </div>
-        <div class="print-style">
-          <span data-type="print-style">ccap-calculus</span>
-        </div>
-        <div class="permissions">
-          <p class="copyright">
-            <span id="copyright-holder-1#{id_suffix}"><a>OSCRiceUniversity</a></span>
-          </p>
-          <p class="license">
-            <a>CC BY</a>
-          </p>
-        </div>
-        <div itemprop="about" data-type="subject">Math</div>
-      </div>
-    HTML
+      HTML
+    )
   end
 end